Get rid of jQuery
Not only that, but just change around styles in JS, so there's no need
for reloading the page, and thus no need for cookies either. We can
generate the font selections by cloning a template element, instead of
handmaking HTML strings.

We can also use Parcel to compile the Less files, as well as inline
fonts.json.

Resolves: #19