Update M PLUS fonts
The author of this wonderful font has made updates to the font as he is
working with Google for inclusion in fonts.google.com.  As part of that
work, the name has also changed from "M+" to "M PLUS".  This change now
contains the modern and new M PLUS fonts.

https://mplusfonts.github.io

Thanks to Coji Morishita for such a beautiful font.